+++ src/Domain/Models/Output/QueryBuilderHelperFile.php
@@ -5,13 +5,14 @@ namespace Soyhuce\NextIdeHelper\Domain\Models\Output;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Support\Collection;
 use ReflectionClass;
+use Soyhuce\NextIdeHelper\Contracts\Amender;
 use Soyhuce\NextIdeHelper\Domain\Models\Entities\Model;
 use Soyhuce\NextIdeHelper\Domain\Models\Entities\QueryBuilder;
 use Soyhuce\NextIdeHelper\Entities\Method;
 use Soyhuce\NextIdeHelper\Support\Output\IdeHelperClass;
 use Soyhuce\NextIdeHelper\Support\Output\IdeHelperFile;
 
-class QueryBuilderHelperFile
+class QueryBuilderHelperFile implements Amender
 {
     private Model $model;